ScriptableObject now used for saving Handles settings
Switched over to a ScriptableObject to save the settings of the Handles, with a single editorpref linking to the path where the asset is stored to make sure one exists.

Also added next/previous point buttons to make switching easier, also realized this is handy to get to the point behind the gameobject instead of having to move them directly.